Throttle speed limitation (#274)
* nidec pas tuning 2

* hardcoded tune

* Fix PTS parameters

* move hardcoded location

* not working - vibrations

* working PAS torque

* working PAS torque

* bug fixes

* working PAS demo

* code cleanup

* Renamed foldback to better represent it's function

* Initial working version

* Added additional comments

* updated comment

* foldback testing

* Basic speed control working, adding apt control

* Started fixing the new module

* Revert on powertrain changes that were accidently kept

* Initial cleanup of merged code

* Working but still testing with velec

* Added comments and remove old tabs

* Fixed PAS bug

* Added new parameter and associated comments

* Removed parameters that hindered vehicle control speed foldback

* Apadted function to enw wheel module and added comments

* Fixed various things base don PR review

---------

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/middleware_vehicle/wheel_speed_sensor.h
@@ -70,7 +70,7 @@ uint32_t WheelSpdSensor_GetSpeedFreq(WheelSpeedSensorHandle_t* pHandle);
   @param  WheelSpeedSensorHandle_t handle
   @return Speed in rpm
 */
-int32_t WheelSpdSensor_GetSpeedRPM(WheelSpeedSensorHandle_t* pHandle);
+uint16_t WheelSpdSensor_GetSpeedRPM(WheelSpeedSensorHandle_t* pHandle);
 
 /**
   @brief  Function to check if speed is detected
